reorg: move project selector out of standalone folder
