+++ apps/gitness/src/components-v2/project-dropdown.tsx
@@ -2,9 +2,9 @@ import { useNavigate, useParams } from 'react-router-dom'
 
 import { DropdownMenu, Icon, Text } from '@harnessio/ui/components'
 
-import { useAppContext } from '../../../framework/context/AppContext'
-import { useRoutes } from '../../../framework/context/NavigationContext'
-import { PathParams } from '../../../RouteDefinitions'
+import { useAppContext } from '../framework/context/AppContext'
+import { useRoutes } from '../framework/context/NavigationContext'
+import { PathParams } from '../RouteDefinitions'
 
 function ProjectDropdown(): JSX.Element {
   const routes = useRoutes()
